+++ spec/eve_online/esi/character_blueprints_spec.rb
@@ -9,7 +9,7 @@ describe EveOnline::ESI::CharacterBlueprints do
 
   specify { expect(subject).to be_a(EveOnline::ESI::Base) }
 
-  specify { expect(described_class::API_PATH).to eq("/v2/characters/%<character_id>s/blueprints/") }
+  specify { expect(described_class::API_PATH).to eq("/v3/characters/%<character_id>s/blueprints/") }
 
   describe "#initialize" do
     context "without options" do
@@ -90,7 +90,7 @@ describe EveOnline::ESI::CharacterBlueprints do
 
   describe "#path" do
     specify do
-      expect(subject.path).to eq("/v2/characters/12345678/blueprints/")
+      expect(subject.path).to eq("/v3/characters/12345678/blueprints/")
     end
   end
 
@@ -102,7 +102,7 @@ describe EveOnline::ESI::CharacterBlueprints do
 
   describe "#url" do
     specify do
-      expect(subject.url).to eq("https://esi.evetech.net/v2/characters/12345678/blueprints/?page=1")
+      expect(subject.url).to eq("https://esi.evetech.net/v3/characters/12345678/blueprints/?page=1")
     end
   end
 end