+++ spec/features/universe_categories_spec.rb
@@ -11,11 +11,11 @@ describe "Get item categories" do
 
   specify { expect(subject.scope).to eq(nil) }
 
-  specify { expect(subject.category_ids.size).to eq(43) }
+  specify { expect(subject.category_ids.size).to eq(47) }
 
   specify { expect(subject.category_ids.first).to eq(0) }
 
   specify { expect(subject.error_limit_remain).to eq(100) }
 
-  specify { expect(subject.error_limit_reset).to eq(13) }
+  specify { expect(subject.error_limit_reset).to eq(15) }
 end