+++ spec/features/universe_group_spec.rb
@@ -7,7 +7,7 @@ describe "Get item group information" do
 
   after { VCR.eject_cassette }
 
-  let(:options) { {id: 450, language: "en-us"} }
+  let(:options) { {group_id: 450, language: "en-us"} }
 
   subject { EveOnline::ESI::UniverseGroup.new(options) }
 
@@ -15,16 +15,16 @@ describe "Get item group information" do
 
   specify do
     expect(subject.as_json).to eq(category_id: 25,
-      group_id: 450,
+      id: 450,
       name: "Arkonor",
       published: true)
   end
 
-  specify { expect(subject.type_ids.size).to eq(10) }
+  specify { expect(subject.type_ids.size).to eq(14) }
 
   specify { expect(subject.type_ids.first).to eq(22) }
 
   specify { expect(subject.error_limit_remain).to eq(100) }
 
-  specify { expect(subject.error_limit_reset).to eq(26) }
+  specify { expect(subject.error_limit_reset).to eq(11) }
 end